pdf error handeling

+++ reporter.py
@@ -4,13 +4,18 @@ from extractors.pdf import extract_pdf
 
 def main(filepath, formatType):
   filepath = Path(filepath)
-  if filepath.suffix == ".pdf":
+  if filepath.suffix.lower() == ".pdf":
     pdfReporter(filepath, formatType)
+  elif filepath.suffix.lower() in [".png", ".jpeg", ".jpg", ".webp", ".gif"]:
+    imageReporter(filepath, formatType)
 
 
 
 def pdfReporter(filepath, formatType):
     data = extract_pdf(filepath)
+    if data is None:
+      print("Could not extract data")
+      return
     if formatType == "text":
       if os.path.exists("PdfReport.txt"):
         os.remove("PdfReport.txt")
@@ -18,4 +23,9 @@ def pdfReporter(filepath, formatType):
         for key, value in data.items():
           f.write(f"{key}: {value} \n")
       f.close()
-        
+    elif formatType == "json":
+      with open("PdfReport.txt", "a") as f:
+        f.write(str(data))
+
+def imageReporter(filepath, formatType):
+  return
